requirements.txt
pandas>=1.3.0
PyYAML>=5.4.0
openai>=1.0.0
python-dotenv>=1.0.0